Checks for Star register table
HccCfg stores pointers to the m_registerSet in m_registerMap.
If the vector is resized then m_registerMap might no longer point
to a valid address in m_registerSet.

Note that ABC has already been adjusted so that it doesn't
use pointers, so doesn't need the checks (so removed here).