Basic feedback from the RawData Processors
The processors get a ClipBoard<struct {integer trigger_tag}> to feedback
the info about data parsing-processing to LoopActions.

LoopActions that inherit the class RecieverOfRawDataProcessingFeedback
can connect the map to such Clipboards.

The struct, the clipboard and the Receiver class are declared in FeedbackBase.h.

The ScanHelper keeps such a map, and sets it in buildScan for the ScanFactory : ScanBase,
and in buildRawDataProcs for the FE processors.

	modified:   src/libYarr/ScanFactory.cpp
	modified:   src/libYarr/ScanHelper.cpp
	modified:   src/libYarr/include/DataProcessor.h
	modified:   src/libYarr/include/FeedbackBase.h
	modified:   src/libYarr/include/ScanFactory.h
	modified:   src/libYarr/include/ScanHelper.h