And shared_ptr for the DataProcessingFeedback ClipboardMap
In principle, during the scan the DataProcessing Feedback ClipboardMap
is owned only one receiver of the feedback, StdDataLoop. But,
I cannot use unique_ptr, because the ScanHelper sets up the receiver
and the senders of feedback (data processors), and then connects them.
I.e. it has to own the Clipboard map, in order to connect
the senders with the receiver during the config & set up time.

	modified:   src/libYarr/ScanFactory.cpp
	modified:   src/libYarr/ScanHelper.cpp
	modified:   src/libYarr/include/ScanFactory.h
	modified:   src/libYarr/include/ScanHelper.h