DataProcessing Feedback Clipboards in FrontEnds
The FronEnd contains `ClipBoard<FeedbackProcessingInfo> clipProcFeedback`
object, which is accessible via Bookkeeper in StdDataLoop and ScanHelper.

ScanHelper connect-s DataProcessors to feedback Clipboards with pointers.
And StdDataLoop accesses the Clipboards via Bookkeeper.

If the feedback clipboard is not used in a DataProcessor, StdDataLoop
waits on the clipboard for `m_dataProcessingTime` which comes from
`average_data_processing_time_us` in the json config.

StdDataLoop ends the iteration when it has received all the expected
triggers `n_triggers_to_receive` on all channels, or when the maximum
total iteration time has passed, `m_totalIterationTime` which comes from
`total_iteration_time_us` in the json config.

	modified:   src/libYarr/include/FrontEnd.h
	modified:   src/libYarr/include/DataProcessor.h
	modified:   src/libStar/include/StarDataProcessor.h
	modified:   src/libYarr/include/FeedbackBase.h
	modified:   src/libYarr/ScanFactory.cpp
	modified:   src/libYarr/ScanHelper.cpp
	modified:   src/libYarr/StdDataLoop.cpp
	modified:   src/libYarr/include/ScanFactory.h
	modified:   src/libYarr/include/ScanHelper.h
	modified:   src/libYarr/include/StdDataLoop.h