Sorted out the loop in header includes
there was a loop in the header includes when DataProcessor.h contained
the struct FeedbackProcessingInfo declaration:
FrontEnd.h -> DataProcessor.h -> ScanBase.h ->
-> LoopEngine -> LoopAction -> Bookkeeper.h -> FrontEnd.h

And it looks like ScanBase.h is actually not used in DataProcessor.

But DataProcessor does include FrontEnd.h to get FrontEndCfg. So, since
both FrontEnd and DataProcessor need the FeedbackProcessingInfo
definition, it should be done in FrontEnd.h (if without extern).

	modified:   src/libYarr/include/DataProcessor.h
	modified:   src/libYarr/include/FrontEnd.h
	modified:   src/libYarr/include/ScanHelper.h