Reverted to getWaitTime, returned maxIterationTime to StdDataLoop
The meaning of getWaitTime is the latency of the HwController.
The HwControllers have it implemented, for example FelixClient and
Netio.

So, unless I missed something, there should be no getReadDelay now.

On the other hand, the timeout limit of the StdDataLoop iteration should
be a parameter of StdDataLoop. So, I reverted it to that. This time is
not a delay, it is a timeout limit, that ends an iteration even if not
all expected triggers are received (which happens for Strips frequently).

	modified:   src/libYarr/StdDataGatherer.cpp
	modified:   src/libYarr/StdDataLoop.cpp
	modified:   src/libYarr/include/RxCore.h
	modified:   src/libYarr/include/StdDataLoop.h