Removed getReadInterval, changed tabs to spaces
getReadInterval seems needed in StdDataLoop, because it spams too much
the debugging log when there is a lost trigger and StdDataLoop has to
loop until the maxIterationTime passes. But it is not necessary, and the
spamming printout should just have a lower logging level, like TRACE.

Changed the printouts in StdDataLoop such that the "no data" messages go
to TRACE of the log.

	modified:   src/libNetioHW/NetioRxCore.cpp
	modified:   src/libYarr/StdDataGatherer.cpp
	modified:   src/libYarr/StdDataLoop.cpp
	modified:   src/libYarr/include/RxCore.h