Moved averageDataProcessingTime back to StdDataLoop
First of all, this is not a delay, it is a timeout. So, it does not
delay anything really, it only adjusts how long StdDataLoop waits for
the feedback from DataProcessors.

In principle, it should be set in the DataProcessors, and there should
be a method `getFeedback` that encapsulates the access to the feedback
Clipboards, instead of accessing them directly in StdDataLoop. But it
would be a lot of work to implement, and it is not clear what exactly
should be implemented. As it is now, it's natural to make it a parameter
of StdDataLoop.

	modified:   src/libNetioHW/NetioRxCore.cpp
	modified:   src/libYarr/StdDataLoop.cpp
	modified:   src/libYarr/include/RxCore.h
	modified:   src/libYarr/include/StdDataLoop.h