With updated ABC mappings, need to adjust some IDs
The emulator test needs to send commands to an ABC by ID, and then
check the response, which is from the HCC so is keyed by input channel.

In addition to changing the IDs, the packets are now sent back in
reverse order (NB this is because the assumed mapping for HCC v0 is
reversed).